requirements.txt
# Vidorag-GraphRAG 双引擎融合 + Flask
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
dashscope
openai>=1.0.0
tqdm
scikit-learn>=1.3.0
torch>=2.0.0
torchvision
# 解决 colpali_engine 与 Qwen2VLProcessor 的兼容性冲突（image_token_id no setter）
transformers==5.3.0
tokenizers==0.22.2
huggingface-hub==1.7.2
qwen-vl-utils
colpali-engine==0.3.14
llama-index-core>=0.12.0
llama-index-instrumentation==0.5.0
llama-index-embeddings-huggingface>=0.4.0
llama-index-readers-file>=0.4.0
defusedxml>=0.7.1
aiosqlite
banks>=2.3.0,<3
tiktoken>=0.7.0
numpy
Pillow
opencv-python-headles
pdf2image
networkx>=3.0
pyarrow>=14.0.0
lancedb
pydantic>=2.0.0
opencv-python-headless>=4.9.0.80
alibabacloud-oss-v2
devtools>=0.12.2
nest_asyncio2
aiofiles
graspologic-native
litellm
azure-identity
spacy
textblob
json-repair
pandas>=2.2
fastdeploy
volcengine>=1.0.0
rouge